third_party/WebKit/LayoutTests/fast/performance/performance-measure-exceptions-expected.txt

This tests that 'performance.measure' throws exceptions with reasonable messages.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.performance.measure("DoesNotExist", "AlsoDoesNotExist") threw exception SyntaxError: The mark 'AlsoDoesNotExist' does not exist..
PASS window.performance.measure("measuring", 'unloadEventStart', "mark") threw exception InvalidAccessError: 'unloadEventStart' is empty: either the event hasn't happened yet, or it would provide cross-origin timing information..
PASS window.performance.measure("measuring", 'unloadEventEnd', "mark") threw exception InvalidAccessError: 'unloadEventEnd' is empty: either the event hasn't happened yet, or it would provide cross-origin timing information..
PASS window.performance.measure("measuring", 'redirectStart', "mark") threw exception InvalidAccessError: 'redirectStart' is empty: either the event hasn't happened yet, or it would provide cross-origin timing information..
PASS window.performance.measure("measuring", 'redirectEnd', "mark") threw exception InvalidAccessError: 'redirectEnd' is empty: either the event hasn't happened yet, or it would provide cross-origin timing information..
PASS window.performance.measure("measuring", 'secureConnectionStart', "mark") threw exception InvalidAccessError: 'secureConnectionStart' is empty: either the event hasn't happened yet, or it would provide cross-origin timing information..
PASS window.performance.measure("measuring", 'domInteractive', "mark") threw exception InvalidAccessError: 'domInteractive' is empty: either the event hasn't happened yet, or it would provide cross-origin timing information..
PASS window.performance.measure("measuring", 'domContentLoadedEventStart', "mark") threw exception InvalidAccessError: 'domContentLoadedEventStart' is empty: either the event hasn't happened yet, or it would provide cross-origin timing information..
PASS window.performance.measure("measuring", 'domContentLoadedEventEnd', "mark") threw exception InvalidAccessError: 'domContentLoadedEventEnd' is empty: either the event hasn't happened yet, or it would provide cross-origin timing information..
PASS window.performance.measure("measuring", 'domComplete', "mark") threw exception InvalidAccessError: 'domComplete' is empty: either the event hasn't happened yet, or it would provide cross-origin timing information..
PASS window.performance.measure("measuring", 'loadEventStart', "mark") threw exception InvalidAccessError: 'loadEventStart' is empty: either the event hasn't happened yet, or it would provide cross-origin timing information..
PASS window.performance.measure("measuring", 'loadEventEnd', "mark") threw exception InvalidAccessError: 'loadEventEnd' is empty: either the event hasn't happened yet, or it would provide cross-origin timing information..
PASS successfullyParsed is true

TEST COMPLETE
